requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
aws-requests-auth==0.4.3
beautifulsoup4==4.14.3
boto3==1.42.59
botocore==1.42.59
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.5
dill==0.4.1
docstring_parser==0.17.0
frozenlist==1.8.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib_metadata==8.7.1
jmespath==1.1.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
markdown-it-py==4.0.0
markdownify==1.2.2
mcp==1.26.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
ollama==0.6.1
opentelemetry-api==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-threading==0.60b1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
packaging==26.0
pillow==12.1.1
prompt_toolkit==3.0.52
propcache==0.4.1
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
referencing==0.37.0
requests==2.32.5
rich==14.3.3
rpds-py==0.30.0
ruff==0.15.4
s3transfer==0.16.0
six==1.17.0
slack_bolt==1.27.0
slack_sdk==3.40.1
soupsieve==2.8.3
sse-starlette==3.3.2
starlette==0.52.1
strands-agents==1.28.0
strands-agents-tools==0.2.21
sympy==1.14.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.41.0
watchdog==6.0.0
wcwidth==0.6.0
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
